requirements.txt
alembic==1.4.3
asyncpg==0.21.0
fastapi==0.62.0
psycopg2==2.8.6
pydantic==1.7.3
pyyaml==5.3.1
SQLAlchemy==1.3.20